Move data_prep/utils.py to llmfoundry
